WorkFlow Tracker
• Developed a website in ASP.net MVC, Javascript, JQuery, HTML and CSS
• There are two roles- user and admin. A new user can register to the
website to use Workflow Tracker.
• Users can create multiple projects, which will automatically create
steps, which are predefined by Admin. User can check status, progress,
actual/estimated start and end date and other details of project and
steps. Users and Admin can edit the details. Admin can enable or disable
users and steps as required.
• They can add comments to mention the changes they made in the steps
with a facility to upload the file.
• There is a facility for the users to generate history report of all
the changes they made on a project or step level. It also provides a
calendar view on a project level, displaying all of the steps on it.